Update domain-to-plan message on the overview page
