+++ client/my-sites/plans-features-main/components/plan-notice-domain-to-plan-credit.tsx
@@ -1,8 +1,13 @@
+import { localizeUrl } from '@automattic/i18n-utils';
+import { useTranslate, formatCurrency } from 'i18n-calypso';
 import QuerySitePlans from 'calypso/components/data/query-site-plans';
-import DomainToPlanCreditMessage from 'calypso/components/domain-to-plan-credit-message';
+import InlineSupportLink from 'calypso/components/inline-support-link';
 import Notice from 'calypso/components/notice';
 import { useDomainToPlanCreditsApplicable } from 'calypso/my-sites/plans-features-main/hooks/use-domain-to-plan-credits-applicable';
+import { useSelector } from 'calypso/state';
+import { getCurrentUserCurrencyCode } from 'calypso/state/currency-code/selectors';
 import type { PlanSlug } from '@automattic/calypso-products';
+
 type Props = {
 	className?: string;
 	onDismissClick?: () => void;
@@ -18,6 +23,11 @@ const PlanNoticeDomainToPlanCredit = ( {
 }: Props ) => {
 	const domainToPlanCreditsApplicable = useDomainToPlanCreditsApplicable( siteId, visiblePlans );
 	const showNotice = domainToPlanCreditsApplicable !== null && domainToPlanCreditsApplicable > 0;
+	const translate = useTranslate();
+	const currencyCode = useSelector( getCurrentUserCurrencyCode );
+	const upgradeCreditDocsUrl = localizeUrl(
+		'https://wordpress.com/support/manage-purchases/upgrade-your-plan/#upgrade-credit'
+	);
 
 	return (
 		<>
@@ -31,7 +41,22 @@ const PlanNoticeDomainToPlanCredit = ( {
 					status="is-success"
 					theme="light"
 				>
-					<DomainToPlanCreditMessage amount={ domainToPlanCreditsApplicable } />
+					{ translate(
+						'You have {{b}}%(amountInCurrency)s{{/b}} in {{a}}upgrade credits{{/a}} available from your current domain. This credit will be applied to the pricing below at checkout if you purchase a plan today!',
+						{
+							args: {
+								amountInCurrency: formatCurrency(
+									domainToPlanCreditsApplicable,
+									currencyCode ?? '',
+									{ isSmallestUnit: true }
+								),
+							},
+							components: {
+								b: <strong />,
+								a: <InlineSupportLink supportLink={ upgradeCreditDocsUrl } />,
+							},
+						}
+					) }
 				</Notice>
 			) }
 		</>
