+++ client/sites/overview/components/plan-credit-notice.tsx
@@ -1,9 +1,12 @@
+import { localizeUrl } from '@automattic/i18n-utils';
 import { Notice } from '@wordpress/components';
+import { useTranslate, formatCurrency } from 'i18n-calypso';
 import QuerySitePlans from 'calypso/components/data/query-site-plans';
 import QuerySitePurchases from 'calypso/components/data/query-site-purchases';
-import DomainToPlanCreditMessage from 'calypso/components/domain-to-plan-credit-message';
+import InlineSupportLink from 'calypso/components/inline-support-link';
 import { useDomainToPlanCreditsApplicable } from 'calypso/my-sites/plans-features-main/hooks/use-domain-to-plan-credits-applicable';
 import { useSelector } from 'calypso/state';
+import { getCurrentUserCurrencyCode } from 'calypso/state/currency-code/selectors';
 import { getSelectedSite } from 'calypso/state/ui/selectors';
 
 const PlanCreditNotice = () => {
@@ -11,6 +14,11 @@ const PlanCreditNotice = () => {
 	const { ID: siteId } = site || {};
 	const domainToPlanCreditsApplicable = useDomainToPlanCreditsApplicable( siteId );
 	const showNotice = domainToPlanCreditsApplicable !== null && domainToPlanCreditsApplicable > 0;
+	const translate = useTranslate();
+	const currencyCode = useSelector( getCurrentUserCurrencyCode );
+	const upgradeCreditDocsUrl = localizeUrl(
+		'https://wordpress.com/support/manage-purchases/upgrade-your-plan/#upgrade-credit'
+	);
 
 	if ( ! siteId ) {
 		return null;
@@ -27,7 +35,22 @@ const PlanCreditNotice = () => {
 					status="info"
 					onRemove={ () => {} }
 				>
-					<DomainToPlanCreditMessage amount={ domainToPlanCreditsApplicable } />
+					{ translate(
+						'You have {{b}}%(amountInCurrency)s{{/b}} in {{a}}upgrade credits{{/a}} available from your current domain. This credit will be applied at checkout if you purchase a plan today!',
+						{
+							args: {
+								amountInCurrency: formatCurrency(
+									domainToPlanCreditsApplicable,
+									currencyCode ?? '',
+									{ isSmallestUnit: true }
+								),
+							},
+							components: {
+								b: <strong />,
+								a: <InlineSupportLink supportLink={ upgradeCreditDocsUrl } />,
+							},
+						}
+					) }
 				</Notice>
 			) }
 		</>
